fix: userAddress in default error pages
