+++ apps/main/src/lend/components/PageVault/Page.tsx
@@ -20,8 +20,8 @@ import { useCurve } from '@evm-ui/features/connect-wallet'
 import { useUserProfileStore } from '@evm-ui/features/user-profile'
 import { useParams } from '@evm-ui/hooks/router'
 import { useMarketMobileFormDrawer, useNewLlamaMarketDetailPage } from '@evm-ui/hooks/useFeatureFlags'
-import { EvmErrorPage } from '@evm-ui/pages/ErrorPage'
-import { MarketType, MarketRateType } from '@evm-ui/types/market'
+import { ErrorPage } from '@evm-ui/pages/ErrorPage'
+import { MarketRateType, MarketType } from '@evm-ui/types/market'
 import { DetailPageLayout } from '@evm-ui/widgets/DetailPageLayout/DetailPageLayout'
 import { DetailPageSection as MarketSection } from '@evm-ui/widgets/DetailPageLayout/DetailPageSection'
 import { t } from '@ui/lib/i18n'
@@ -58,11 +58,12 @@ export const Page = () => {
 
   const error = marketError ?? apiMarket.error
   return error ? (
-    <EvmErrorPage
+    <ErrorPage
       title={t`Error`}
       subtitle={error.message}
       error={error}
       continueUrl={getCollateralListPathname(params)}
+      userAddress={userAddress}
     />
   ) : (
     <MarketContextProvider
