+++ packages/evm-ui/src/shared/ui/DataTable/DataTable.tsx
@@ -2,6 +2,7 @@
 import { type ReactNode, useMemo, useRef } from 'react'
 import { useLayoutStore } from '@evm-ui/features/layout'
 import { TablePagination } from '@evm-ui/shared/ui/DataTable/TablePagination'
+import { EvmErrorMessage } from '@evm-ui/shared/ui/EvmErrorMessage'
 import Box from '@mui/material/Box'
 import { Theme } from '@mui/material/styles'
 import Table from '@mui/material/Table'
@@ -16,7 +17,6 @@ import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useIsMobile } from '@ui/hooks/useBreakpoints'
 import { t } from '@ui/lib/i18n'
 import { EmptyStateCard, EmptyStateCardProps } from '../EmptyStateCard'
-import { EvmErrorMessage } from '../ErrorMessage'
 import { DATA_TABLE_CATEGORIES, type DataTableCategory, type DataTableCategoryConfig } from './categories'
 import { DataTableHeaderHeight, type useCurveTable } from './data-table.utils'
 import { DataRow, type DataRowProps } from './DataRow'
