+++ packages/evm-ui/src/shared/ui/EvmErrorMessage.tsx
@@ -1,0 +1,7 @@
+import type { ComponentProps } from 'react'
+import { useConnection } from 'wagmi'
+import { ErrorMessage } from '@evm-ui/shared/ui/ErrorMessage'
+
+export const EvmErrorMessage = (props: Omit<ComponentProps<typeof ErrorMessage>, 'userAddress'>) => (
+  <ErrorMessage {...props} userAddress={useConnection().address} />
+)
